Refactor presentation package structure.
