+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name="com.voltek.newsfeed.presentation.splash.SplashActivity"
+            android:name="com.voltek.newsfeed.presentation.ui.splash.SplashActivity"
             android:theme="@style/SplashTheme">
 
             <intent-filter>
@@ -26,16 +26,16 @@
         </activity>
 
         <activity
-            android:name="com.voltek.newsfeed.presentation.list.ListActivity"
+            android:name="com.voltek.newsfeed.presentation.ui.list.ListActivity"
             android:launchMode="singleTask" />
 
         <activity
-            android:name="com.voltek.newsfeed.presentation.details.DetailsActivity"
-            android:parentActivityName="com.voltek.newsfeed.presentation.list.ListActivity" />
+            android:name="com.voltek.newsfeed.presentation.ui.details.DetailsActivity"
+            android:parentActivityName="com.voltek.newsfeed.presentation.ui.list.ListActivity" />
 
         <activity
-            android:name="com.voltek.newsfeed.presentation.news_sources.NewsSourcesActivity"
-            android:parentActivityName="com.voltek.newsfeed.presentation.list.ListActivity" />
+            android:name="com.voltek.newsfeed.presentation.ui.news_sources.NewsSourcesActivity"
+            android:parentActivityName="com.voltek.newsfeed.presentation.ui.list.ListActivity" />
 
     </application>
