+++ app/src/main/java/com/voltek/newsfeed/data/Provider.kt
@@ -1,25 +1,12 @@
 package com.voltek.newsfeed.data
 
 import android.support.annotation.IdRes
-import com.voltek.newsfeed.data.entity.ArticleRAW
 import com.voltek.newsfeed.data.entity.SourceRAW
-import io.reactivex.Single
 
 /**
  * Set of interfaces, that provides access to data layer
  */
 object Provider {
-
-    object Api {
-
-        interface Articles {
-            fun get(source: String): Single<List<ArticleRAW>>
-        }
-
-        interface NewsSources {
-            fun get(): Single<List<SourceRAW>>
-        }
-    }
 
     object Storage {
