+++ app/src/main/java/com/voltek/newsfeed/presentation/base/BaseActivity.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.presentation
+package com.voltek.newsfeed.presentation.base
 
 import android.content.Context
 import android.content.Intent
