+++ app/src/main/java/com/voltek/newsfeed/presentation/base/BaseModel.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.presentation
+package com.voltek.newsfeed.presentation.base
 
 /**
  * Base ViewModel class.
