+++ app/src/main/java/com/voltek/newsfeed/presentation/entity/ArticleUI.kt
@@ -1,4 +1,4 @@
-package com.voltek.newsfeed.domain.entity
+package com.voltek.newsfeed.presentation.entity
 
 import org.parceler.Parcel
