--- app/src/main/java/com/voltek/newsfeed/presentation/splash/SplashView.kt
@@ -1,5 +1,0 @@
-package com.voltek.newsfeed.presentation.splash
-
-import com.voltek.newsfeed.presentation.BaseView
-
-interface SplashView : BaseView
