+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/details/DetailsActivity.kt
@@ -1,12 +1,12 @@
-package com.voltek.newsfeed.presentation.details
+package com.voltek.newsfeed.presentation.ui.details
 
 import android.os.Bundle
 import com.voltek.newsfeed.R
-import com.voltek.newsfeed.domain.entity.ArticleUI
+import com.voltek.newsfeed.presentation.entity.ArticleUI
 import com.voltek.newsfeed.navigation.command.CommandOpenWebsite
 import com.voltek.newsfeed.navigation.command.CommandShareArticle
 import com.voltek.newsfeed.navigation.proxy.Command
-import com.voltek.newsfeed.presentation.BaseActivity
+import com.voltek.newsfeed.presentation.base.BaseActivity
 import org.parceler.Parcels
 
 class DetailsActivity : BaseActivity() {
