+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/details/DetailsContract.kt
@@ -1,9 +1,9 @@
-package com.voltek.newsfeed.presentation.details
+package com.voltek.newsfeed.presentation.ui.details
 
 import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType
-import com.voltek.newsfeed.presentation.BaseModel
-import com.voltek.newsfeed.presentation.BaseView
+import com.voltek.newsfeed.presentation.base.BaseModel
+import com.voltek.newsfeed.presentation.base.BaseView
 
 object DetailsContract {
