+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/list/ListActivity.kt
@@ -1,17 +1,17 @@
-package com.voltek.newsfeed.presentation.list
+package com.voltek.newsfeed.presentation.ui.list
 
 import android.content.Intent
 import com.voltek.newsfeed.R
-import com.voltek.newsfeed.domain.entity.ArticleUI
+import com.voltek.newsfeed.presentation.entity.ArticleUI
 import com.voltek.newsfeed.navigation.command.CommandOpenArticleDetailsScreen
 import com.voltek.newsfeed.navigation.command.CommandOpenNewsSourcesScreen
 import com.voltek.newsfeed.navigation.command.CommandOpenWebsite
 import com.voltek.newsfeed.navigation.proxy.Command
-import com.voltek.newsfeed.presentation.BaseActivity
+import com.voltek.newsfeed.presentation.base.BaseActivity
 import com.voltek.newsfeed.navigation.command.CommandShareArticle
-import com.voltek.newsfeed.presentation.details.DetailsActivity
-import com.voltek.newsfeed.presentation.details.DetailsFragment
-import com.voltek.newsfeed.presentation.news_sources.NewsSourcesActivity
+import com.voltek.newsfeed.presentation.ui.details.DetailsActivity
+import com.voltek.newsfeed.presentation.ui.details.DetailsFragment
+import com.voltek.newsfeed.presentation.ui.news_sources.NewsSourcesActivity
 import kotlinx.android.synthetic.main.activity_list.*
 import org.parceler.Parcels
