+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/news_sources/NewsSourcesContract.kt
@@ -1,11 +1,11 @@
-package com.voltek.newsfeed.presentation.news_sources
+package com.voltek.newsfeed.presentation.ui.news_sources
 
 import com.arellomobile.mvp.viewstate.strategy.AddToEndSingleStrategy
 import com.arellomobile.mvp.viewstate.strategy.StateStrategyType
 import com.voltek.newsfeed.R
-import com.voltek.newsfeed.domain.entity.SourceUI
-import com.voltek.newsfeed.presentation.BaseModel
-import com.voltek.newsfeed.presentation.BaseView
+import com.voltek.newsfeed.presentation.entity.SourceUI
+import com.voltek.newsfeed.presentation.base.BaseModel
+import com.voltek.newsfeed.presentation.base.BaseView
 
 object NewsSourcesContract {
