+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/splash/SplashActivity.kt
@@ -1,13 +1,13 @@
-package com.voltek.newsfeed.presentation.splash
+package com.voltek.newsfeed.presentation.ui.splash
 
 import android.content.Intent
 import com.arellomobile.mvp.presenter.InjectPresenter
 import com.voltek.newsfeed.navigation.command.CommandOpenArticlesListScreen
 import com.voltek.newsfeed.navigation.command.CommandOpenNewsSourcesScreen
 import com.voltek.newsfeed.navigation.proxy.Command
-import com.voltek.newsfeed.presentation.BaseActivity
-import com.voltek.newsfeed.presentation.list.ListActivity
-import com.voltek.newsfeed.presentation.news_sources.NewsSourcesActivity
+import com.voltek.newsfeed.presentation.base.BaseActivity
+import com.voltek.newsfeed.presentation.ui.list.ListActivity
+import com.voltek.newsfeed.presentation.ui.news_sources.NewsSourcesActivity
 
 class SplashActivity : BaseActivity(),
         SplashView {
