+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/splash/SplashPresenter.kt
@@ -1,17 +1,17 @@
-package com.voltek.newsfeed.presentation.splash
+package com.voltek.newsfeed.presentation.ui.splash
 
 import com.arellomobile.mvp.InjectViewState
 import com.voltek.newsfeed.NewsApp
 import com.voltek.newsfeed.domain.exception.NoNewsSourcesSelectedException
-import com.voltek.newsfeed.domain.entity.SourceUI
+import com.voltek.newsfeed.presentation.entity.SourceUI
 import com.voltek.newsfeed.domain.interactor.Parameter
 import com.voltek.newsfeed.domain.interactor.news_sources.NewsSourcesInteractor
 import com.voltek.newsfeed.navigation.command.CommandOpenArticlesListScreen
 import com.voltek.newsfeed.navigation.command.CommandOpenNewsSourcesScreen
 import com.voltek.newsfeed.navigation.command.CommandSystemMessage
 import com.voltek.newsfeed.navigation.proxy.Router
-import com.voltek.newsfeed.presentation.BasePresenter
-import com.voltek.newsfeed.presentation.Event
+import com.voltek.newsfeed.presentation.base.BasePresenter
+import com.voltek.newsfeed.presentation.base.Event
 import io.reactivex.functions.Action
 import io.reactivex.functions.Consumer
 import javax.inject.Inject
