+++ app/src/main/java/com/voltek/newsfeed/presentation/ui/splash/SplashView.kt
@@ -1,0 +1,5 @@
+package com.voltek.newsfeed.presentation.ui.splash
+
+import com.voltek.newsfeed.presentation.base.BaseView
+
+interface SplashView : BaseView
